requirements.txt
pandas
numpy
scikit-learn
streamlit
Pillow
google-generativeai